requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.5
django-countries==7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
flake8==3.8.4
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.5
sqlparse==0.4.1
toml==0.10.2
